requirements.txt
dj-database-url
Django
django-ckeditor
django-heroku
django-highlightjs
django-js-asset
django-taggit
gunicorn
psycopg2
Pillow
whitenoise